UltimateOutdoorgen4: Added Tai's progHouse Tileset/map
The door event for going out the house gets you stuck. I tried to changed the location but it did nothing. You have to CTRL to get out. So please fix this after pulling this commit.

The event in the beginning works fine when walking out and talking to Ceiba and Mom.