Fixed Screen Tone for Cutscenes
Genius

Also made it night time when you can access Cerigold, so we can have the festival at night. Will need to do some serious work to redo Cerigold cutscenes so the festival is lively and lights/candles are giving off light, but I know it's worth it. Taking a break though :)