Catch and Store fix
When replacing a pokemon in your party with the one you just caught, it will now be in the slot that the pokemon you sent to the box was in. So if you swap out your party leader, the pokemon you just added to the party will now be in the front.