hotheaded / coolheaded / healing sun
simple and unremarkable abilities, yet had too specific names for some reason.
Great abilities are ones that describe general traits, such as being a glutton, being competitive or describe how the pokemon is (Ice Body, Dry Skin, Keen Eye, etc), thats the reasoning for the renaming
~~if you want some hyper specific shit you have "zealous dance"~~

also, made the healing during sun thing but as a different ability, name was taken from Pokemon Master's EX but without the number attached to it as this version doesn't have "tiers" of healing.