slow start rework / low mode stuff
---------- Slow Start ----------
On the first switch-in, this Pokemon's Attack and Speed are weakened by 33% for 5 turns. These turns burnt don't reset even if switched out.

PS: could have reworked this to be a omniboost after the 5 turns but Merry preferred this version;
---------- Honorbound ----------
The user cannot switch out when below half health, this check is made whenever the user is taken below half health (like emergency exit) or when the user switches in (taking in account hazards).
Currently, not even shed shell can overwrite this effect. (Prone to change)

---- THIS IS FOR LOW MODE ONLY ----
gave honorbound to the pseudos wannabe birds
jack said he wanted to give the sirens liquid voice and its a pretty mediocre ability so whatever (btw thanks Chesp for the melumourn rename)
melumander had 5 more bst than the other sirens, fixed that