Paid the power bill
Fixed issue with the screen not fading back in when exiting the player's house after the initial Ceiba and Mom convo outside.
For some reason RPG Maker just decided it wasn't going to process a command to change a self switch after transferring the player. Literally nothing changed to cause this that I can think of or find.
This was fixed by putting the "self switch A on" command before the player transfer event rather than after the transfer event