Several things
Got a burn timer set up
The mixture will burn every 5 seconds, but that timer will reset if you complete a stir
The stir direction arrows will change when you complete a stir. It's random if it's the same direction or not
Stage 1 ends when a timer runs out, and that is set to 30 seconds right now
I'm thinking of adding some type of reward for the more stirs you complete, like a higher chance of pokeblock quality, but ever so slightly an increase

Next is adding berries I guess? And then changing the mixture hue to match the resulting pokeblock